Extracts post list component.
