+++ .vitepress/theme/components/HomeSections.module.css
@@ -5,178 +5,3 @@
 .recentPostsSection {
   padding: 0.75rem 0.5rem 0.5rem;
 }
-
-.recentPostsList {
-  display: flex;
-  flex-direction: column;
-}
-
-.recentPostsItem {
-  padding: 0;
-}
-
-.recentPostsItem + .recentPostsItem {
-  border-top: 1px solid rgba(148, 163, 184, 0.18);
-}
-
-.recentPostsLink {
-  position: relative;
-  display: grid;
-  grid-template-columns: minmax(0, 180px) minmax(0, 1fr) auto;
-  gap: 1.5rem;
-  align-items: start;
-  padding: 1.15rem 0.15rem;
-  color: inherit;
-  text-decoration: none;
-}
-
-.recentPostsLink::before {
-  content: "";
-  position: absolute;
-  left: -0.5rem;
-  top: 0.6rem;
-  bottom: 0.6rem;
-  width: 3px;
-  border-radius: 999px;
-  background: linear-gradient(180deg, #60a5fa, #f97316);
-  opacity: 0;
-  transform: scaleY(0.35);
-  transform-origin: center;
-  transition:
-    opacity 0.24s ease,
-    transform 0.24s ease;
-}
-
-.recentPostsItem:hover .recentPostsLink::before,
-.recentPostsLink:focus-visible::before {
-  opacity: 1;
-  transform: scaleY(1);
-}
-
-.recentPostsItem:hover .recentPostsArrow,
-.recentPostsLink:focus-visible .recentPostsArrow {
-  opacity: 1;
-  transform: translateX(0);
-}
-
-.recentPostsLink:focus-visible {
-  outline: 2px solid rgba(147, 197, 253, 0.8);
-  outline-offset: 2px;
-}
-
-.recentPostsMeta {
-  padding-top: 0.25rem;
-  color: #93c5fd;
-  font-size: 0.88rem;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-}
-
-.recentPostsBody {
-  min-width: 0;
-}
-
-.recentPostsTitle {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.55rem;
-  margin: 0;
-  color: #fff;
-  font-family: "Lato", sans-serif;
-  font-size: clamp(1.25rem, 1.8vw, 1.6rem);
-  transition:
-    transform 0.24s ease,
-    color 0.24s ease;
-}
-
-.recentPostsVisited {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 1.15rem;
-  height: 1.15rem;
-  border-radius: 999px;
-  background: rgba(22, 163, 74, 0.22);
-  color: #4ade80;
-  box-shadow: 0 0 0 1px rgba(74, 222, 128, 0.35);
-}
-
-.recentPostsVisited svg {
-  width: 0.85rem;
-  height: 0.85rem;
-}
-
-.recentPostsExcerpt {
-  margin: 0.7rem 0 0;
-  color: var(--vp-c-text-2);
-  line-height: 1.8;
-  transition:
-    transform 0.24s ease,
-    color 0.24s ease;
-}
-
-.recentPostsItem:hover .recentPostsTitle,
-.recentPostsLink:focus-visible .recentPostsTitle {
-  transform: translateX(4px);
-  color: #dbeafe;
-}
-
-.recentPostsItem:hover .recentPostsExcerpt,
-.recentPostsLink:focus-visible .recentPostsExcerpt {
-  transform: translateX(4px);
-  color: var(--vp-c-text-1);
-}
-
-.recentPostsArrow {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 2.4rem;
-  height: 2.4rem;
-  align-self: center;
-  border: 1px solid rgba(148, 163, 184, 0.2);
-  border-radius: 9999px;
-  background: rgba(15, 23, 42, 0.35);
-  color: #dbeafe;
-  opacity: 0;
-  transform: translateX(10px);
-  transition:
-    opacity 0.22s ease,
-    transform 0.22s ease,
-    border-color 0.22s ease,
-    background 0.22s ease;
-}
-
-.recentPostsItem:hover .recentPostsArrow {
-  border-color: rgba(147, 197, 253, 0.4);
-  background: rgba(30, 41, 59, 0.7);
-}
-
-.recentPostsArrow svg {
-  width: 1.2rem;
-  height: 1.2rem;
-}
-
-@media (max-width: 960px) {
-  .recentPostsLink {
-    grid-template-columns: 1fr;
-    gap: 0.75rem;
-    padding: 1rem 0.05rem;
-  }
-
-  .recentPostsArrow {
-    justify-self: start;
-    transform: translateX(0);
-    opacity: 1;
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .recentPostsLink::before,
-  .recentPostsTitle,
-  .recentPostsExcerpt,
-  .recentPostsArrow {
-    transition: none;
-  }
-}
